requirements.txt
dash==2.14.2
dash_bootstrap_components==1.5.0
joblib==1.3.2
kafka_python==2.0.2
pandas==1.4.3
plotly==5.11.0
PyMySQL==1.1.0
SQLAlchemy==1.4.7
scikit-learn==1.2.2 
numpy
gunicorn
python-dotenv
cryptography
psutil